code_reviewer/requirements.txt
requests==2.31.0
click==8.1.7
colorama==0.4.6
dataclasses==0.6; python_version<'3.7'
openai==1.3.0
azure-identity==1.15.0
azure-core==1.29.5
